Added checking win condition (fully functional) and game over screen.
- Tested out win condition checking many times and it works
consistently. The code is a bit of a mess considering how difficult it
is to abstract much out (there probably is a way, just needs more time
and analysis).
- Game over screen is the bare minimum in terms of looks; it works and
that's basically it.
- The game can handle either player winning or a draw event and can
reset the board afterward. Deciding which player should go first during
the next game has not been implemented (it will always start with player
1).
- I was messing around with discord's rich presence integration with
eclipse using a plugin some guy made (in beta as of now) and now there's
an additional file in .settings for that. Rich presence still isn't
working.